Add files via upload
This project is still in progress. There is much more I want to add. There will be a quiz game that will be exactly like my final year project for the blind. But the game will be more attractive and interactive because the main focus of my project is for the child's heart patients. Those pictures is several interface of my mobile app for IJN.